test flyway config still don't works
