+++ src/main/java/com/cargoHub/dto/OrderResponseDto.java
@@ -1,4 +1,4 @@
-package com.cargohub.dto.jar;
+package com.cargohub.dto;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
