--- src/main/java/com/cargoHub/dto/jar/DeliveryStatus.java
@@ -1,7 +1,0 @@
-package com.cargohub.dto.jar;
-
-public enum DeliveryStatus {
-    ON_THE_WAY,
-    DELIVERED,
-    PROCESSING,
-}
