+++ src/main/java/com/cargohub/entities/Dimensions.java
@@ -1,5 +1,6 @@
 package com.cargohub.entities;
 
+import com.cargohub.entities.transports.CarrierCompartment;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -27,12 +28,15 @@ public class Dimensions {
     @Column
     Integer length;
 
+    @OneToOne(mappedBy = "volume", orphanRemoval = true, cascade = CascadeType.ALL)
+    CarrierCompartment carrierCompartment;
+
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
         Dimensions that = (Dimensions) o;
-        return  width.equals(that.width) &&
+        return width.equals(that.width) &&
                 height.equals(that.height) &&
                 length.equals(that.length);
     }
