+++ src/main/resources/application.properties
@@ -10,6 +10,8 @@ spring.jpa.show-sql=true
 spring.flyway.url=jdbc:mysql://localhost:3306/cargo_transports?serverTimezone=UTC
 spring.flyway.user=root
 spring.flyway.password=CargoUsers
+spring.flyway.locations=db/migration
+spring.flyway.ignore-future-migrations=true
 
 # do not change
 spring.data.neo4j.username=neo4j
