+++ src/test/java/com/cargohub/service/impl/CargoPositionServiceImplTest.java
@@ -1,8 +1,8 @@
-import com.cargohub.entities.Dimensions;
-import com.cargohub.entities.transports.CarrierCompartment;
-import com.cargohub.exceptions.CarrierCompartmentException;
-import com.cargohub.repository.CarrierCompartmentRepository;
-import com.cargohub.service.impl.CarrierCompartmentServiceImpl;
+package com.cargohub.service.impl;
+
+import com.cargohub.entities.CargoPosition;
+import com.cargohub.exceptions.CargoPositionException;
+import com.cargohub.repository.CargoPositionRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.ArgumentCaptor;
@@ -25,26 +25,23 @@ import static org.mockito.ArgumentMatchers.nullable;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-class CarrierCompartmentServiceImplTest {
+class CargoPositionServiceImplTest {
 
     @InjectMocks
-    CarrierCompartmentServiceImpl service;
+    CargoPositionServiceImpl service;
 
     @Mock
-    CarrierCompartmentRepository repository;
+    CargoPositionRepository repository;
 
-    CarrierCompartment subject;
-    Page<CarrierCompartment> page;
+    CargoPosition subject;
+    Page<CargoPosition> page;
     Pageable pageable;
 
     @BeforeEach
     void setUp() {
         MockitoAnnotations.initMocks(this);
-        subject = new CarrierCompartment();
+        subject = new CargoPosition();
         subject.setId(9992);
-        subject.setVolume(new Dimensions());
-        subject.setMaximumWeight(999d);
-        subject.setFreeSpace(.35);
 
         page = new PageImpl(List.of(subject));
         pageable = PageRequest.of(0, 10);
@@ -53,25 +50,25 @@ class CarrierCompartmentServiceImplTest {
     @Test
     void create() {
         subject.setId(null);
-        when(repository.save(nullable(CarrierCompartment.class))).thenReturn(subject);
-        ArgumentCaptor<CarrierCompartment> captor = ArgumentCaptor.forClass(CarrierCompartment.class);
-        CarrierCompartment returned = service.save(subject);
+        when(repository.save(nullable(CargoPosition.class))).thenReturn(subject);
+        ArgumentCaptor<CargoPosition> captor = ArgumentCaptor.forClass(CargoPosition.class);
+        CargoPosition returned = service.save(subject);
         verify(repository).save(captor.capture());
-        CarrierCompartment used = captor.getValue();
+        CargoPosition used = captor.getValue();
         assertThat(used, is(subject));
         assertThat(returned, is(subject));
     }
 
     @Test
     void createThrowsCargoPositionException() {
-        assertThrows(CarrierCompartmentException.class, () -> service.save(subject));
+        assertThrows(CargoPositionException.class, () -> service.save(subject));
     }
 
     @Test
     void findById() {
         when(repository.findById(nullable(Integer.class))).thenReturn(Optional.of(subject));
         ArgumentCaptor<Integer> captor = ArgumentCaptor.forClass(Integer.class);
-        CarrierCompartment returned = service.findById(subject.getId());
+        CargoPosition returned = service.findById(subject.getId());
         verify(repository).findById(captor.capture());
         Integer used = captor.getValue();
         assertThat(used, is(subject.getId()));
@@ -81,28 +78,28 @@ class CarrierCompartmentServiceImplTest {
     @Test
     void findByIdThrowsCargoPositionException() {
         when(repository.findById(nullable(Integer.class))).thenReturn(Optional.empty());
-        assertThrows(CarrierCompartmentException.class, () -> service.findById(subject.getId()));
+        assertThrows(CargoPositionException.class, () -> service.findById(subject.getId()));
         subject.setId(null);
-        assertThrows(CarrierCompartmentException.class, () -> service.findById(subject.getId()));
+        assertThrows(CargoPositionException.class, () -> service.findById(subject.getId()));
     }
 
     @Test
     void update() {
-        when(repository.save(nullable(CarrierCompartment.class))).thenReturn(subject);
+        when(repository.save(nullable(CargoPosition.class))).thenReturn(subject);
         when(repository.existsById(nullable(Integer.class))).thenReturn(true);
-        ArgumentCaptor<CarrierCompartment> captor = ArgumentCaptor.forClass(CarrierCompartment.class);
-        CarrierCompartment returned = service.update(subject);
+        ArgumentCaptor<CargoPosition> captor = ArgumentCaptor.forClass(CargoPosition.class);
+        CargoPosition returned = service.update(subject);
         verify(repository).save(captor.capture());
-        CarrierCompartment used = captor.getValue();
+        CargoPosition used = captor.getValue();
         assertThat(used, is(subject));
         assertThat(returned, is(subject));
     }
 
     @Test
     void updateThrowsCargoPositionException() {
-        assertThrows(CarrierCompartmentException.class, () -> service.update(subject));
+        assertThrows(CargoPositionException.class, () -> service.update(subject));
         subject.setId(null);
-        assertThrows(CarrierCompartmentException.class, () -> service.update(subject));
+        assertThrows(CargoPositionException.class, () -> service.update(subject));
     }
 
 
@@ -118,14 +115,14 @@ class CarrierCompartmentServiceImplTest {
 
     @Test
     void deleteThrowsCargoPositionException() {
-        assertThrows(CarrierCompartmentException.class, () -> service.delete(subject.getId()));
+        assertThrows(CargoPositionException.class, () -> service.delete(subject.getId()));
     }
 
     @Test
     void findAll() {
         when(repository.findAll(any(Pageable.class))).thenReturn(page);
         ArgumentCaptor<Pageable> captor = ArgumentCaptor.forClass(Pageable.class);
-        Page<CarrierCompartment> returned = service.findAll(pageable);
+        Page<CargoPosition> returned = service.findAll(pageable);
         verify(repository).findAll(captor.capture());
         Pageable used = captor.getValue();
         assertThat(returned, is(page));
