+++ src/test/java/com/cargohub/service/impl/CargoServiceImplTest.java
@@ -1,10 +1,10 @@
+package com.cargohub.service.impl;
 
 import com.cargohub.entities.Cargo;
 import com.cargohub.entities.CargoPosition;
 import com.cargohub.entities.enums.DeliveryStatus;
 import com.cargohub.exceptions.CargoException;
 import com.cargohub.repository.CargoRepository;
-import com.cargohub.service.impl.CargoServiceImpl;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.*;
