+++ src/test/java/com/cargohub/service/impl/CarrierCompartmentServiceImplTest.java
@@ -1,7 +1,9 @@
+package com.cargohub.service.impl;
+
 import com.cargohub.entities.Dimensions;
-import com.cargohub.exceptions.DimensionsException;
-import com.cargohub.repository.DimensionsRepository;
-import com.cargohub.service.impl.DimensionsServiceImpl;
+import com.cargohub.entities.transports.CarrierCompartment;
+import com.cargohub.exceptions.CarrierCompartmentException;
+import com.cargohub.repository.CarrierCompartmentRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.ArgumentCaptor;
@@ -24,23 +26,26 @@ import static org.mockito.ArgumentMatchers.nullable;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-class DimensionsServiceImplTest {
+class CarrierCompartmentServiceImplTest {
 
     @InjectMocks
-    DimensionsServiceImpl service;
+    CarrierCompartmentServiceImpl service;
 
     @Mock
-    DimensionsRepository repository;
+    CarrierCompartmentRepository repository;
 
-    Dimensions subject;
-    Page<Dimensions> page;
+    CarrierCompartment subject;
+    Page<CarrierCompartment> page;
     Pageable pageable;
 
     @BeforeEach
     void setUp() {
         MockitoAnnotations.initMocks(this);
-        subject = new Dimensions();
+        subject = new CarrierCompartment();
         subject.setId(9992);
+        subject.setVolume(new Dimensions());
+        subject.setMaximumWeight(999d);
+        subject.setFreeSpace(.35);
 
         page = new PageImpl(List.of(subject));
         pageable = PageRequest.of(0, 10);
@@ -49,25 +54,25 @@ class DimensionsServiceImplTest {
     @Test
     void create() {
         subject.setId(null);
-        when(repository.save(nullable(Dimensions.class))).thenReturn(subject);
-        ArgumentCaptor<Dimensions> captor = ArgumentCaptor.forClass(Dimensions.class);
-        Dimensions returned = service.save(subject);
+        when(repository.save(nullable(CarrierCompartment.class))).thenReturn(subject);
+        ArgumentCaptor<CarrierCompartment> captor = ArgumentCaptor.forClass(CarrierCompartment.class);
+        CarrierCompartment returned = service.save(subject);
         verify(repository).save(captor.capture());
-        Dimensions used = captor.getValue();
+        CarrierCompartment used = captor.getValue();
         assertThat(used, is(subject));
         assertThat(returned, is(subject));
     }
 
     @Test
     void createThrowsCargoPositionException() {
-        assertThrows(DimensionsException.class, () -> service.save(subject));
+        assertThrows(CarrierCompartmentException.class, () -> service.save(subject));
     }
 
     @Test
     void findById() {
         when(repository.findById(nullable(Integer.class))).thenReturn(Optional.of(subject));
         ArgumentCaptor<Integer> captor = ArgumentCaptor.forClass(Integer.class);
-        Dimensions returned = service.findById(subject.getId());
+        CarrierCompartment returned = service.findById(subject.getId());
         verify(repository).findById(captor.capture());
         Integer used = captor.getValue();
         assertThat(used, is(subject.getId()));
@@ -77,28 +82,28 @@ class DimensionsServiceImplTest {
     @Test
     void findByIdThrowsCargoPositionException() {
         when(repository.findById(nullable(Integer.class))).thenReturn(Optional.empty());
-        assertThrows(DimensionsException.class, () -> service.findById(subject.getId()));
+        assertThrows(CarrierCompartmentException.class, () -> service.findById(subject.getId()));
         subject.setId(null);
-        assertThrows(DimensionsException.class, () -> service.findById(subject.getId()));
+        assertThrows(CarrierCompartmentException.class, () -> service.findById(subject.getId()));
     }
 
     @Test
     void update() {
-        when(repository.save(nullable(Dimensions.class))).thenReturn(subject);
+        when(repository.save(nullable(CarrierCompartment.class))).thenReturn(subject);
         when(repository.existsById(nullable(Integer.class))).thenReturn(true);
-        ArgumentCaptor<Dimensions> captor = ArgumentCaptor.forClass(Dimensions.class);
-        Dimensions returned = service.update(subject);
+        ArgumentCaptor<CarrierCompartment> captor = ArgumentCaptor.forClass(CarrierCompartment.class);
+        CarrierCompartment returned = service.update(subject);
         verify(repository).save(captor.capture());
-        Dimensions used = captor.getValue();
+        CarrierCompartment used = captor.getValue();
         assertThat(used, is(subject));
         assertThat(returned, is(subject));
     }
 
     @Test
     void updateThrowsCargoPositionException() {
-        assertThrows(DimensionsException.class, () -> service.update(subject));
+        assertThrows(CarrierCompartmentException.class, () -> service.update(subject));
         subject.setId(null);
-        assertThrows(DimensionsException.class, () -> service.update(subject));
+        assertThrows(CarrierCompartmentException.class, () -> service.update(subject));
     }
 
 
@@ -114,14 +119,14 @@ class DimensionsServiceImplTest {
 
     @Test
     void deleteThrowsCargoPositionException() {
-        assertThrows(DimensionsException.class, () -> service.delete(subject.getId()));
+        assertThrows(CarrierCompartmentException.class, () -> service.delete(subject.getId()));
     }
 
     @Test
     void findAll() {
         when(repository.findAll(any(Pageable.class))).thenReturn(page);
         ArgumentCaptor<Pageable> captor = ArgumentCaptor.forClass(Pageable.class);
-        Page<Dimensions> returned = service.findAll(pageable);
+        Page<CarrierCompartment> returned = service.findAll(pageable);
         verify(repository).findAll(captor.capture());
         Pageable used = captor.getValue();
         assertThat(returned, is(page));
