+++ src/test/resources/application.properties
@@ -1,0 +1,17 @@
+server.port=9041
+
+spring.datasource.url=jdbc:h2:mem:myDb;DB_CLOSE_DELAY=-1
+spring.datasource.username=sa
+spring.datasource.password=""
+
+#spring.jpa.hibernate.ddl-auto=update
+spring.jpa.show-sql=true
+
+spring.flyway.url=jdbc:h2:mem:myDb;DB_CLOSE_DELAY=-1
+spring.flyway.user=sa
+spring.flyway.password=""
+spring.flyway.locations=db/test/migration
+
+# do not change
+spring.data.neo4j.username=neo4j
+spring.data.neo4j.password=Graph1
